ub:ubfi:skipped address of subtable 0 in ubrt

Created-by: qiushengming
Commit-by: Yuhao Xiang;Xiongchuan Zhou;Anzhe Li;Yahui Liu;Junlong Zheng;Jianquan Lin
Merged-by: slim6882
Description: ub:ubus: Add ummu_map attribute in sysfs
ub:ubus: Fix X86 build error
dma-mapping: benchmark: add support for UB devices
ub:ubus: call ub_host_probe inside register_ub_manage_subsystem_ops
ub:ubus: Support for UB Memory Decoder general layer implementation
...

See merge request: kunpengcompute/anolis-cloud-kernel!67

+++ drivers/ub/ubus/services/hotplug/hotplug_core.c
@@ -8,6 +8,7 @@
 #include "../../enum.h"
 #include "../../route.h"
 #include "../../port.h"
+#include "../../task.h"
 #include "../service.h"
 #include "hotplug.h"
 
@@ -478,7 +479,7 @@ static int ubhp_handle_link_up(struct ub_slot *slot)
 		goto err_link_up;
 	}
 
-	ret = ub_enum_entities_active(&dev_list);
+	ret = ub_enum_entities_active(&dev_list, TASK_SRC_SELF);
 	if (ret) {
 		ub_err(slot->uent, "hotplug start devices failed, ret=%d\n", ret);
 		goto err_link_up;
@@ -636,7 +637,7 @@ static void ubhp_handle_present(struct ub_slot *slot)
 		return;
 	}
 
-	if (ubhp_handle_link_up(slot))
+	if (!slot->r_uent && ubhp_handle_link_up(slot))
 		goto poweroff;
 
 	ubhp_set_indicators(slot, INDICATOR_ON, INDICATOR_NOOP);
